Added IR tracers associated with the column package
I added several new tracers to the suite of tracers advected by IR.
This requires some additional calls to subroutine add_tracer_to_tracer_linked_list at initialization:

    call add_tracer_to_tracer_linked_list(tracersHead, 'iceAge',               'iceVolumeCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'firstYearIceArea',     'iceAreaCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'levelIceArea',         'iceAreaCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'levelIceVolume',       'iceVolumeCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'pondArea',             'levelIceArea')
    call add_tracer_to_tracer_linked_list(tracersHead, 'pondDepth',            'pondArea')
    call add_tracer_to_tracer_linked_list(tracersHead, 'pondLidThickness',     'pondArea')
    call add_tracer_to_tracer_linked_list(tracersHead, 'snowScatteringAerosol','snowVolumeCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'iceScatteringAerosol', 'iceVolumeCategory')
    call add_tracer_to_tracer_linked_list(tracersHead, 'iceBodyAerosol',       'iceVolumeCategory')

In these calls, the first quantity in quotes is the tracer and the second is its parent.
These tracers are added to the tracer linked list depending on the values of the relevant
logical options for packages (pkgColumnTracerIceAgeActive, pkgColumnTracerLevelIceActive, etc.).

To support these new tracers, I added a number of fields to the Registry.
For example, the iceAge tracer is associated with new fields iceAgeMask, iceAgeEdgeFlux, iceAgeProduct,
iceAgeTriangle, iceAgeBarycenterx, iceAgeBarycentery, iceAgeCenter, iceAgeGradx, iceAgeGrady,
iceAgeCons, iceAgeLocalMin and iceAgeLocalMax.  And similarly for the other new tracers.

Note that pondDepth and pondLidThickness have 3 parents.  This is the first time IR has been used to
advect tracers with 3 parents in MPAS CICE, and the first time using the new barycenter formulas for such tracers.
(Barycenter formulas for tracers with > 2 parents are not included in standard CICE.)

To check that the formulas are working correctly, I wrote a new subroutine called recoverTracerMeans.
This subroutine does an exact integral of mass*tracer products based on the reconstructed linear
fields (as defined by the center, gradx and grady values for each tracer) and compares it to
the product of the mean values of mass*tracer.  I verified that for tracers with any number of parents <= 3,
the reconstructed fields recover the mean to within roundoff, as expected.

The recoverTracerMeans subroutine includes embedded tracer loops and is not very efficient.
However, this subroutine is used only for testing, and the call is turned off by default.