Update MPAS component drivers to validate time steps
This commit adds validation to model init that the coupling interval for
the component is sensibly configured with the model time step.

For example, if the time step is longer than a coupling interval, the
model errors out. Additionally, if the time steps selected would not
land on a coupling interval boundary, the model also errors out.