Init snow thickness to max in glacier land units in CLM
Previously snow thickness is initialized to half of the maximum value so
as not to bias ablation or accumulation in the SMB field coming out of
CLM.  However, this meant that it tooks months of integration before the
coupler and the GLC model would see a nonzero SMB.  This made tests slow
- the standard PE setup required the test to run for close to an hour
before nonzero SMB appeared over a substantial portion of an ice sheet.

This changes sets the initial snow thickness to the max value so that
positive SMB will be generated immediately.  Note that negative SMB will
take some (~months) to develop since the entire snow thickness will have
to be ablated first.  However, since our primary focus in ACME is
Antarctica where most of the ice sheet has positive SMB year-round, it
makes more sense to see +SMB immediately.