Pass fields used in atm_compute_dyn_tend( ) as arguments into new routine atm_compute_dyn_tend_work( )
The intent of this commit is to present the fields used in the computational parts of
atm_compute_dyn_tend( ) as simple array dummy arguments, rather than as the pointers that
are queried from pools. The dummy arguments are declared as assumed shape arrays
in atm_compute_dyn_tend_work( ).