Set isActive=.true. for the field used to smooth coordinate surfaces in init_atmosphere core
We create a field "on the fly" in the real-data initialization case for MPAS-Atmosphere
so that we can properly apply smoothers to coordinate surfaces in parallel. The original
code was written before the "isActive" component was added to field types. The "isActive"
component will be required by future development (specifically, to ensure that halos
are only exchanged for active fields), so we need to set isActive to .true. before we can
exchange halos for this field.