Set isActive=.true. for temporary fields created in the atmosphere core
There are two places in the atmosphere core where we create a field "on the fly": one
in the monotonic advection routine, and another in the physics interface when projecting
horizontal wind tendencies to edges; in both cases, the purpose is to allow us to exchange
halos for temporary fields. The original code that uses these fields was written before
the "isActive" component was added to field types. The "isActive" component will be required
by future development (specifically, to ensure that halos are only exchanged for active fields),
so we need to set isActive to .true. before we can exchange halos for these fields.