Changed sub_ice_shelf_2D to use ocn_init_ssh_and_ssp
The sub_ice_shelf_2D test case now supports iterative initialization
of SSP to be consistent with the desired SSH.

For the first iteration, the setup should include:
&init_ssh_and_ssp
    config_read_ssh_and_ssp_from_stream = .false.
/
<stream name="sea_surface_iteration"
	type='none'
	filename_template='none'
	input_interval="initial_only" >

	<var name="seaSurfacePressure"/>
	<var name="ssh"/>
</stream>
The initial guess at the SSP is computed as follows:
1. Compute ssh and reference grid (restingThickness, refBottomDepth, etc.) as before
2. Compute reference T and S based on a z* vertical grid with ssh=0
3. Call ocn_init_ssp_from_ssh to compute the initial guess at the SSP (computed to match bottom pressure)

For iterations beyond the first, the following namelists and streams should be used:
&init_ssh_and_ssp
    config_read_ssh_and_ssp_from_stream = .true.
/
<stream name="sea_surface_iteration"
        type='input'
        filename_template=<path/to/output/from/previous/forward/run>
        input_interval="initial_only" >

        <var name="seaSurfacePressure"/>
        <var name="ssh"/>
</stream>
A new SSH will be read in from a short forward run and the SSP will be adjusted to try to reduce the change in SSH in subsequent iterations.

The salinity calculation now uses zMid computed for a z-star
coordinate.  This will allow the addition of bottom topography
in the future.