Allow inner dimensions to be defined at run-time or as constants at compile-time
For optimal performance, -DCONST_INNER_DIMS can be added to the definition of MODEL_FORMULATION
in the top-level Makefile; when doing so, the following additional variables need to be defined
appropriately in MODEL_FORMULATION: CONST_NVERTLEVELS, CONST_MAXEDGES, CONST_MAXEDGES2, and
CONST_NUM_SCALARS.

If CONST_INNER_DIMS is not defined in MODEL_FORMULATION, values for the inner dimensions
nVertLevels, maxEdges, maxEdges2, and num_scalars will based on the input file at run-time.

Defining the inner dimensions as constants at compile-time can significantly improve performance
of the dynamics routines. However, if inner dimensions are defined as constants, and those constants
don't match the dimensions found in the input file, the code will halt with a fatal error.