Adding a bootstrap routine for analysis members
This commit adds a bootstrap routine to the analysis member driver which
can be used to perform setup of things like streams before they are
read, which happens before analysis members are initialized.

An example of where this is used is the time_series_stats analysis
member, which currently builds a restart stream during initialization,
however the restart stream needs to be built before it's read. So, the
constructon of the restart stream now occurs during bootstrapping,
rather than initialization.