Switched index order in IR subroutine define_local_to_global_transformations
The subroutine that creates matrices for transforming local to global transformations
has some vector and matrix arrays with dimension (3,3).  Previously, the second index
contained the x, y and z Cartesian coordinates of a given vector.  I switched indices
so that the x, y and z Cartesian coordinates are now contained in the first (inner) index.
This prevents some warning messages associated with memory strides.

Answers are BFB.