Add mask for cells albany will extend to, add special case for albany vertex in MISMIP
This is needed to properly calculate which vertices (triangles) should
be used by Albany.  In particular, there is a special case of 1
Dirichlet dynamic cell, 1 Dirichlet nondynamic cell, and 1 extended cell
that should be made an active vertex (triangle) for MISMIP test case to
work correctly.  This commit also adds the logic to check for that
situation.