Read input streams at the beginning of init mode
No mutable streams were being read at the beginning of init mode.
With this change, mutable input streams are read after the
immutable streams.  This will allow, for example, an iterative
initialization procedure in which fields from a (probably very
short) forward run are used as input to the next run in init mode.
Such an iterative procedure can be used to initialize the SSP and
SSH to be in balance.