Have interface reconstruct velo at all edges, let MPAS decide which to use
MPAS only 'keeps' normalVelocity on edges that are defined as 'dynamic'
in edgeMask.

Now that normalVelocity is being calculated on all edges of the FEM
mesh, special logic is needed for the MPAS edge locations that
correspond to the edges of the FEM mesh.

For boundary edges one of the two triangles sharing the edge is
not part of the velocity solver's FEM mesh, so if the edge is not in
the triangle that does exist, then instead project the edge location
on to the edge of the triangle that exists.

This will be slightly inaccurate in these situations, but this situation
only occurs for edges between two Dirichlet nodes and only in a variable
resolution mesh (and only roughly half of those).  (Note that these
edges will be solved for all the time, but the values are only actually
used by MPAS in the situation where the edge is between two Dirichlet
nodes.)