Added a subroutine to hold the calving front fixed; added simple prescribed velocity options
In ACME v1 we plan to hold the calving front fixed, at least for the first round of experiments.
To support this choice I added a new config option, config_restore_calving_front.
The default is false.  If set to true, the model will call a new subroutine (li_restore_calving_front)
in lieu of the regular calving subroutine (li_calve_ice).

The new subroutine loops through all cells.
For each cell with a bed below sea level, the model checks for two things:

(1) If ice was present initially but now is very thin or absent, the ice is restored to
a small thickness called restoreThicknessMin, which is set to config_dynamic_thickness/10.
The ice that is added is put in a new field, restoreThickness, in order to keep track
of energy non-conservation.

(2) If ice was absent initially but now is present, it is removed and added to calvingThickness.

To support this logic, I added a logical function li_mask_is_initial_ice.

In addition, I created a new module, li_velocity_simple. This module supports two options for prescribing
a simple velocity field:

(1) 'uniform' = uniform velocity in a straight line.  The speed and direction are controlled
by the parameters flowSpeed and flowTheta, set in subroutine li_velocity_simple_block_init.

(2) 'radial' = outward from a central point, with speed increasing linearly with distance from
the center.  The increase of speed with distance is controlled by the parameter flowGradient.
The center of flow is (xCenter,yCenter), set to (0,0) by default.

Both options assume flow on a plane (not on the surface of a sphere).
They can be set in the namelist, with config_velocity_solver = 'simple' and
config_simple_velocity_type = 'uniform' or 'radial'.

To create these simple velocity fields, I prescribed (u,v) at cell centers and then found the
normal velocity components by taking the dot product of (u,v) with the unit normal vector
at each cell midpoint.  I verified that the normal components are correct for a periodic domain.

As a bug check, I reconstructed the center velocity from the normal components.
For the periodic mesh used in the circular shelf problem, the reconstruction works in the mesh interior
but is incorrect along the boundary.  Doug reports that the reconstruction coefficients should be correct
for a v1 compliant periodic mesh, but Matt says that the land-ice periodic meshes are not yet v1 compliant.

Other changes:

- I created a new mask, activeForCalvingMask, to support the calving thickness threshold option.
This option treats grounded ice as active where thickness > config_dynamic_thickness, but treats
floating ice as active where thickness > config_calving_thickness.

- For brevity, I changed 'config_calving_thickness_threshold' to 'config_calving_thickness',
and 'config_calving_topographic_threshold' to 'config_calving_topography'.
The default values are now 100 m and -500 m, respectively.  I checked that it is OK for config_calving_thickness
to be >, < or = to config_dynamic_thickness.

- I decided not to support the config_calving_on_startup option, at least for now.
The issue is that subroutine calve_ice is called with input argument 'domain', which means it cannot
be called from within landice_init_block, which otherwise would be the most natural place to do
calving on startup.  We can add this option later, if needed.

- I modified some logic so that vector reconstruction coefficients are always computed at initialization,
and then are given a halo update. This is in anticipation that we will almost always want these coefficients.
For example, incremental remapping requires them.

- I added scyr = 31536000.0 to the mpas_li_constants module. This is the number of seconds in a 365-day year.
Although MPAS-LI uses SI units, it is useful to have this constant available (e.g., for velocity diagnostics
where it's easier to think in terms of m/yr than m/s).

- To support various test cases, I added some code to li_velocity_simple_init_block to tweak the
thickness, bedTopography and sfcMassBal fields. I will remove this code later but am including it
in this commit, to make it easy to repeat the testing if desired.

Using these simple velocity fields in the circular-shelf test problem, I verified that the calving subroutines
are working as intended.  In particular:

When config_restore_calving_front = .true.:
- With prescribed radially outward flow, ice advances beyond the initial calving front but then is removed.
- With a prescribed negative SMB along the periphery, ice behind the initial calving front melts away,
but then is restored with a thickness of restoreThicknessMin.

When config_calving = 'floating':
- Ice calves wherever the ice is floating (and not elsewhere).

When config_calving = 'topographic_threshold':
- Ice calves wherever the bed topography lies lower than the topographic threshold (and not elsewhere).

When config_calving = 'thickness_threshold':
- Ice calves wherever the ice is floating and is thinner than config_calving_thickness
(apart from a protected one-cell ring around floating ice that is below the thickness threshold).
- The calving front can advance and retreat under appropriate forcing.

When config_calving_timescale > 0:
- Ice calves at the desired rate; a fraction of the ice is removed during each time step.