Add option to disable name mangling in interface
Currently the Fortran to C++ interface for external velocity solvers
uses a series of preprocessor #define's to deal with the different way
that Fortran and C++ use underscores in routine/function names.

However, on some machine configurations it may be required to NOT do
this, namely, on Mira where the Fortran and C++ compilers that must be
used are from different vendors.

This commit adds "#ifndef MPASLI_EXTERNAL_INTERFACE_DISABLE_MANGLING"
around the #define's to provide the possibility for avoiding this.
Currently this variable is only set within ACME and is not used at all
in standalone MPAS.