Added optional ssh and restingThickness arguments
restingThickness is required if ssh is present.

The layerThickness, zMid and the restingThickness with a z*
vertical coordinate depressed to the given ssh.

The layerThickness should be bit-identical to what it was before if
the ssh argument is not present.  zMid will be different to machine
roundoff because it is summed from layerThickness rather than
being computed directly from refBottomDepth as before.