Remove some unnecessary files
I removed the old polygon intersection code that was in the forward_model directory.
This was used during IR development but is no longer needed.

I also removed a setup script written by Elizabeth, which previously was used
to stage files from the column package (before the column package became part
of the repo).

I checked that the code still builds from scratch and gives the same answers for
a standard test case.