bug fix/cleanup: no intra-proc IO MPI commun.
This commit removes communciation from a processor to itself by
modifying the IO halo to ensure it does not include its host processor.

This is conceptually cleaner and rectifies a bug that
did not occur for openmpi/1.8.1 but did for openmpi/1.6.5

The issue was that messages communicated from the host to the host
were corrupted on 1.6.5.  Now, no IO communication occurs from host
to host which increases efficiency and avoids this bug.

This commit also now explicitly addresses the corner case where a
particle on processor A with ioProc A is being sent to processor B.
Before, B would not show up in A's ioProc halo due to restructuring in
1beb979.  Now, this case is explicitly covered in order to ensure
symmetry of the ioProc communication graph which is required for the MPI
communication to occur for IO.