Fixing an issue with reused buffers in li statistics
This commit fixes an issue in the mpas_li_statistics module where
dmpar allreduce (i.e. max_int, min_int, etc...) routines used the same
buffer for the send and receive. Certain MPI implementations fail when
this is done (i.e. MPICH), so this commit changes the behavior to have
independnt buffers for each part of the communication.