clean up and extend the del4 operator for momentum.
add a configure option to scale the div part of the del4 operator relative to the curl part.
new parameter in hmix_del4 namelist record is config_mom_del4_div_factor with a default value of 1.0

other minor clean up:
when possible, use dvEdge (instead of dcEdge/sqrt(3.0)).
when dvEdge is very short, use 0.25 of dcEdge