working time-based particle resets
Tested in config_AM_lagrPartTrack_reset_criteria = 'particle' and
'global_value' modes with scripts at
https://gist.github.com/84d008a138bc427d6ece

Tested with planar periodic grid to make sure particles are reset at the
right frequencies.  A potential optimization would be to specify, based
on a decomposition, the currentCellReset instead of currently hardcoding
it to be -1 in
src/core_ocean/analysis_members/mpas_ocn_lagrangian_particle_tracking_reset.F:270.
This, however, requires more inteligent ways of initializing particles
depending upon runtime condition.  This is essentially an optimization
for performance so the current naive approach is used.

Note that spatial reset information is time invarient.  It would be
possible to allow time-based resets in the future by making the
dimensions to be of size (nParticles, Time), for xParticleReset, etc.
However, this would require an all-to-all communication to process
the halos and this would be quite expensive and break scalability.

Note, there was a need to have a distinction between Eulerian halos for
particle computation and the full Lagrangian computational halo (the
result of the prior refractoring commit).