Fixed an IR mask issue to ensure exact restart
Previously, test cases on the gx1 and gx3 meshes were not restarting exactly.
There were small differences in pondArea and related variables for a few cells.

I traced the problem to the make_masks subroutine, which assigns a value of 1
wherever a tracer value is physically meaningful, and 0 otherwise.
A value of 1 was erroneously being assigned to pondArea for iCell = nCells+1
(i.e., non-existing cells) because the parent tracer, levelIceArea, had been
initialized to 1.0 everywhere (including iCell = nCells+1) for the base run.
As a result, the gradient-limiting calculation for pondArea gave a slightly different
answer than it would have if non-existent neighbors were excluded.

The initialization of levelIceArea to 1.0 everywhere was skipped for config_do_restart = true,
which is why the answers were different on restart.

To fix the problem, I rewrote the masking routine to initially assign a mask value of 0
everywhere (including iCell = nCells+1).  Then there are a series of loops from 1 to nCells
that can set the mask to 1 based on the values of parent tracers, but the mask value
remains 0 for nCells+1.

It was not necessary to change the initialization of levelIceArea. At some point, we might
want to initialize this and other variables in a way that does not assign non-zero values
in non-existent cells.

This commit is answer-changing for cells that previously had the wrong mask values.
I verified that answers are BFB for test problems with column physics turned off.