Updating logic to handle missing unlimited dimensions
This commit updates the logic that is used to determine if an unlimited
dimension exists or not in an input file. Previously, if -1 was returned
as the dimension ID it was assumed that there wasn't an unlimited
dimension ID. This is now updated to check for any negative or zero value.

It was reported by PIO developers that the pio_inquire call should
return PIO_EBADID if the unlimited dimension is missing, but currently
this isn't the case in our tests. At a later point in time this could be
updated to check PIO_EBADID instead of the dimension ID.