Removed the call to mpas_init_reconstruct when using the external FO dycore
In a previous calving commit, I changed some logic in subroutine landice_init_block
such that mpas_init_reconstruct would always be called (in anticipation of needing
reconstruction coefficients for IR transport). This change was premature.
At present, the periodic_hex grid generator does not set the attributes x_period
and y_period.  As a result, mpas_init_reconstruct dies with a seg fault when called
with the periodic MISMIP mesh.  Thanks for Matt Hoffman for help in sorting this out.

With this commit, I went back to the old logic, such that the reconstruction
routine is not called with the external FO dycore (unless certain config
variables are explicitly set to 'true').

I re-ran a circular-shelf test case for calving and verified that the results
have not changed.