Remove wind stress diag variables
The variables
  windStressZonalDiag
  windStressMeridionalDiag
were removed from the code in V3, but were kept in the Registry by
mistake.  They are not connected to anything.

In stand-alone mode, there is currently no way to see zonal and
meridional versions of surface stress.  The natural way to do that now
is to put it in an analysis member.

In ACME, one can look at the variables
  windStressZonal
  windStressMeridional