Re-sequence of high-level coupler flow and copy of necessary GLC files
This commit contains a fairly substantial re-ordering of GLC-relevant
calls in cesm_comp_mod run loop.  Re-ordering was necessary (in
principle and theory) in order to accomodate the new shelf-ocean
coupling, which involves a remapping to the GLC grid, calculation of
boundary fluxes, and a remapping back to the OCN grid, within an OCN
coupling timestep (or less).  The commit builds (after including some
necessary CISM-era GLC files that were missed the last commit).
The latter necessitated a change to the mpasli-based driver Makefile.

This commit builds, but has not been run-tested.