Added a vertical temperature/enthalpy solver and supporting code
I translated a CISM module, glissade_therm.F90, to MPAS data structures.
The new module is called mpas_li_thermal.F.  It has the following public subroutines:

(1) li_thermal_init: This subroutine initializes the temperature profile.
The options (given by config_thermal_init) are 'linear' and 'file'.
- If 'linear', then this subroutine constructs a simple (but somewhat realistic)
linear temperature profile in each column, varying from surfaceAirTemperature at the
upper surface to slightly below the pressure melting point temperature at the bed.
- If 'file', then it is assumed that the temperature has already been read from an input file.
- On restarts, this config setting is ignored and the temperature is read
from the restart file.

(2) li_thermal_solver: This subroutine solves for the evolution of the temperature/
enthalpy profile in each column over one timestep, given heat sources and sinks
in the ice interior and at the bed. It is called at the beginning of the prognostic
timestep, before the transport calculations. The thermal solver operates on all cells
with thickness > config_thermal_thickness (= 1 m by default).

The options (given by config_thermal_solver) are 'none', 'temperature' and 'enthalpy'.
- The default for now is 'none', in which case the thermal solver does nothing.
- The 'temperature' option is functionally equivalent to the Glissade temperature solver,
which has been tested extensively.
- Likewise, the 'enthalpy' option is equivalent to the Glissade enthalpy solver,
which has been tested for long EISMINT-2 simulations and some Greenland simulations,
but not as extensively as the temperature solver. In the enthalpy case,
there is an additional tracer called 'waterfrac'.

(3) li_heat_dissipation_sia: This subroutine computed the heat dissipation rate (deg/s)
in the ice interior, assuming shallow-ice stresses. There is a similar subroutine in Glide.
This subroutine differs from Glide, however, in that the heat dissipation is computed on
cell edges (rather than vertices) before being averaged to cell centers.

There are additional private subroutines with the following functions:

- Compute melting at the bed.

- Compute the pressure melting point temperature in the column and at the bed.

- Given temperature and waterfrac, compute enthalpy; and given enthalpy, compute
temperature and waterfrac.

- Construct tridiagonal matrix elements in each column for the temperature or enthalpy solver,
and solve the matrix.

I added several new fields and config options in the Registry.
Here is the current list of fields in the thermal state:
- temperature
- waterfrac
- enthalpy
- surfaceAirTemperature
- surfaceTemperature
- basalTemperature
- surfaceConductiveFlux
- basalConductiveFlux
- basalHeatFlux
- basalFrictionFlux
- heatDissipation

This code still needs to be tested in some simple cases to verify that the answers
are the same (to a good approximation) as those given by CISM.