Add missing package to bulk forcing fields
This commit adds a missing package to one of the fields used in bulk
forcing. Previously if the bulk thickness flux namelist option was
turned on but the active tracers bulk surface forcing was disabled, the
model would die with a segfault. This allows thickness fluxes without
tracer fluxes, and vice versa.