Adding a module for interpolation in init mode
Contains subroutines for vertical linear interpolation and
horizontal nearest-neighbor and bilinear interpolation

The vertical interpolation is performed on a single column, and
makes no assumptions about the number or placement of input or
output levels.  The routine is primarily intended to be used for
interpolating activeTracers from one vertical grid to another,
but is general enough to be used for any linear interpolation
of a field sampled in the vertical at specified points.  Neither
the input nor the output fields are required ot be of size nVertLevels.

The horizontal interpolation routines assume a logically rectangular
input grid that can optionally be periodic in either direction.
Horizontal interpolation can be performed on 2D (horizontal) or 3D
input data (resulting in 1D or 2D output grids, respectively). For
3D interpolation, the input and output grids are assumed to have the
same number of vertical levels, and maxLevelCell is not used to limit
the range of interpolation for a given vertical column.  The number
of vertical levels need not be nVertLevels.