Implement Burrows et al. (2014) marine organic sea spray emissions.
This commit adds two implementations of marine organic sea spray
emissions, as described in Burrows et al., ACP (2014).  They can be
activated by passing the appropriate -chem option to CAM_CONFIG_OPTS.
The two implementations are as follows:

    * trop_mam9: detailed implementation adding 2 aerosol modes to
      MAM7, for a total of 9 modes, and 3 new chemical species.  This
      option adds a total of 14 advected tracers.

    * trop_mam4_mom: simplified implementation adding a single new
      chemical species in the existing aerosol modes of MAM4.  This
      option adds a total of 4 advected tracers.

This code should not be answer-changing unless the new options are
activated, in which case it will be non-BFB and potentially
climate-changing.  If turned on, the new chemistry options will also
add a number of history variables.

The new code also includes several alternative parameterizations for
marine organic emissions; these are turned off by default and require
code modifications (via a hard-coded switch) to be turned on.

Default paths to needed input data files (available in svn repo) are
set in namelist_defaults.xml, so the code should work out-of-the-box
using default data files without need for additional namelists.

[BFB]
[FCC]
[AG-143]