Add bc and pom in the coarse mode of the MAM3/MAM4 chem pkgs
This commit adds BC and POM in the 3rd mode (coarse) of MAM3 and MAM4
chemistry packages. To run the model is this configuration, user
has to configure the model using "rain_evap_to_coarse_aero" option.
This option will act as a CPP directive where turning this on would
add BC and POM into the coarse mode of MAM3 or MAM4 chemistry
packages.

This commit also fixes a bug in the chemistry preprocessor.

[BFB]