edgeVunpack using receive buffer now
This was the last remaining difference between the CPU and GPU code in terms of implementation. Now, internal values are copied to the receive buffer, and edgeVunpack uses the receive buffer. This helps in two ways. First, there is a lower overall amount of data to transfer. Second, copying internal values has no dependencies. This gave a slight speed-up for GPU code.