Fixed a memory leak in IR
Previously, the IR code was allocating a tracer_type pointer called dummyTracer,
allocating some arrays within the derived type, and then nullifying the pointer
without deallocating the arrays. This resulted in a memory leak.

Now, when the code is done with the dummyTracer pointer, the arrays are deallocated
and then the pointer is deallocated. This fixes the memory leak.

This change is BFB.