Cleaned up the IR zap_small_mass subroutine
The zap_small_mass subroutine is responsible for zeroing out mass (i.e., iceAreaCategory)
and tracers wherever the mass falls below a small threshold.

Previously, this subroutine had loops over the tracer linked list inside a loop over cells.
I rewrote the subroutine so that the cell loops are inside the loop over the linked list.
This is more efficient.

To put the cell loops inside the tracers loops, I allocated a scratch mask array as is done
in MPAS Landice. For this purpose I added a "scratch" var struct to the Registry.

Currently, the full ACME code is seg-faulting in the zap_small_mass subroutine.
My hope is that these changes will fix the seg fault, but that remains to be seen.

Note: When running the full sea ice model, it may not be necessary to call the zap_small_areas
subroutine inside the IR, provided a similar subroutine is called elsewhere.