Adding driver script generation
This commit adds generation for driver scripts for test cases that have
multiple case directory and steps. The driver script can be used to
define the order each step should be performed, and automate the
process of running a test case.

Documentation is added to the README file about how to define driver
scripts.