Updating <get_file> tag to download into a backup location
This commit modifies several tags related to downloading and linking
files. These changes are not backwards compatible.

Now the <get_file> tag has a dest_path and file_name attribute that
define the name of the file that will be downloaded, and where it should
be downloaded into. dest_path can have a value of case, in which case
it's downloaded into the case directory that is created.

Additionally, the local mirror option was removed, as it is no longer
needed. Instead, the <add_link> tag has a source_path option that allows
links to files that exist in paths defined in a config file.

Coupled together, these modifications replace the previous behavior, but
allow files to be symlinked into cases instead of downloaded multiple
times (saving disk space as a result).