Change `kissvec` to be BFB with old version.
In the translation from C, the KISS random number generator was modified
in a way that caused it to provide random floats that differ from the
Fortran version by roundoff-level differences. This tweaks the C version
to give the exact same answers as the original code, at least for
current Intel compilers.

Testing:

    Test suite: Intel and GNU benchmark.
    Baseline: N/A
    Namelist changes: None.
    Status: bit-for-bit (since the code is still not used in CESM)