Adding a summary of OpenMP threads for the ocean core
This commit adds writing of the number of OpenMP threads available to an
MPI task within the log.*.err file.

Note: The thread number is the number available to an MPI task, but the
MPI task may not use them. For example, if there is no parallel region,
no additionaly OpenMP threads will be used.