Fixing an issue with KPP that prevents bit-reproducibility
This commit fixes three issues with KPP. To begin with, it changes two
casts to floats to cast to RKIND reals to ensure the types are
consistent.

Additionally, normvalVelocityAv previously was computed using the
incorrect iEdge (iEdge instead of iEdgeVal). This made KPP incorrect and
non-bit reproducible in most configurations.