restructuring to speed up computation of particles
This refractor removes capability to keep track of halos as they are
modified because this was determined to be an expensive computation,
particularly for a large number of processors because the number of
computational steps are larger than the number of write or restart
steps.

Also rearranged send/recv for less latency: MPI_IRecv should be called
before MPI_ISend so all sends have a "hook" to receive them.