pure refractoring, results should be bit identical
Combination of multiple squashed commits:

1) Tested with LIGHT_planar_periodic_periodic_advection test case.
Note that if there are problems it is likely this commit is to blame
and it should be tested.

Refractoring in anticipation of generalizing the computational
halos to accomodate particle resets.

2) This commit allows for processors to send a particle to
an arbitrary currentBlock via update of the computational
halos.  However, at present the computational halos should
be static in time and this commit should consequently be
identical in computation to the previous commit.

3) generalized computational halos to account for possibility of
evolving halo, generalization uses infrastructure for
IO halos

4) prepartation for particle resets commented out
analagous communication patterns as for IO halos

5) This ensures that halos as currently written are not broken by
considering 'currentBlockReset' as a component in building the halo.
This current commit was tested with all 'currentBlockReset'=0 as well
as 'currentBlockReset'='currentBlock' at initialization time.

6) Tested with 48 processes for planar periodic 20km case
where particles are placed randomly.  IO and computational
halos grow and shrink dynamically to account for variable
communication patterns to be able to communicate with
ioHalo and currentBlockReset.  These tests suggest that
particles should be resetable within the current halo infrastructure.

7) implicit merge of code with bug fix commit (fixup)
2870ba758462575e849d5a8fe635cccb9e3e8405

mpi communication bug fix with > 1000 processors

This fixes an issue where there was corruption of array temporaries
during MPI_ISend statements due to Fortran vs C-array ordering.
Essentially, array temporaries were being formed for the send data.
This did not present an issue on small processor counts and certain
architectures but caused problems that appeared during usage on
Edison with the 30-10km Idealized Southern Ocean.