Adding seaSurfacePressure to the restart stream
This commit fixes an issue when trying to restart the ocean model in a
coupled mode. Previously seaSurfacePressure was not part of the restart
stream, but on restart the coupler may not provide a seaSurfacePressure,
in which case the model is unable to restart exactly.

Additionally, to allow whether seaSurfacePressure is read from the
restart stream or not, a namelist option is added named
'config_restart_forcing_fields'. When this is set to `.false.`
seaSurfacePressure will not be read from a restart stream even if it
exists in the file.