Attempt to get scripting infrastructure to run faster.
Replaces xmlquery and preview_namelists with ACME python versions.

Make xmlquery support retrieval of multiple values at once. Replace
the common meme in CIME of doing multiple xmlqueries with a single
xmlquery call. This will greatly reduce the number of forks and the
number of times the XML files need to be parsed.

Preliminary testing on skybridge shows a 1-2 minute speed-up of
a create_test acme_developer build-only run. When factoring in the parallelism
in create_test, we can assume about 20 minutes of overall compute
time saved.

[BFB]