Added an option to advect tracers along with thickness
Before this commit, the ice thickness was transported with the velocity field,
but tracers such as ice temperature were not advected.  With this commit,
it is possible to advect thickness and tracers together using a first-order
upwind scheme for both.

The new option is activated by setting both config_thickness_advection = 'fo'
and config_tracer_advection = 'fo' in the namelist. The default value of
config_tracer_advection remains 'none', in which case thickness but not
temperature is advected.

When tracer advection is turned on, a new subroutine called li_tendency_thickness_tracers
is called instead of the old subroutine li_tendency_thickness.  The new subroutine
is similar in structure to the old one.  However, it does not compute tracer tendencies;
it simply updates the advected tracers to the new values.  I couldn't find a simple way
to use tracer tendencies correctly without adding an extra time level to the tracer fields.

Upwind advection of temperature and tracers is carried out by a new subroutine
called advect_thickness_tracers_upwind.  I verified that both mass and internal energy
are conserved under horizontal transport.

The new tendency subroutine also calls a vertical_remap subroutine to map tracer values from the
modified layers (after horizontal transport and mass balance) back to the standard layers
(defined by layerInterfaceSigma). I verified that the remapping conserves internal energy.
Rather than use the existing vertical_remap subroutine in mpas_li_diagnostic_vars.F,
I added a slightly modified private subroutine in mpas_li_tendency.F.

Note: The location of the vertical_remap call depends on whether we are advecting thickness
alone, or thickness and tracers together.  If thickness alone (the default), then
the vertical_remap call is in diagnostic_solve_before_velocity.  If thickness
and tracers together, then the vertical_remap call is in the li_tendency module.
It makes more sense to do vertical remapping in the tendency calculation, but
I've left it in place for the default case to avoid breaking tests.
At some point we should have a single vertical_remap subroutine that can be
called from either tendency routine.

The details of the tracer advection depend on the value of config_thermal_solver.
If config_thermal_solver = 'temperature', then temperature is the advected tracer,
and the advection is fairly straightforward. But if config_thermal_solver = 'enthalpy',
enthalpy must be advected in order to conserve energy. For this purpose I added
tracer_setup and tracer_finish subroutines to convert between temperature/waterfrac
and enthalpy. This required adding an 'li_thermal' dependency to the Makefile.

I added a Registry variable called maxTracersAdvect.  This is the maximum number of
tracers to be advected and is currently set to 1 (since it is sufficient to advect
either temperature or enthalpy).

In the li_thermal module, I added some tracer halo updates that were previously missing.
I modified the names and interfaces of the temperature-enthalpy conversion subroutines,
which are now public.

I retained the Registry structure whereby temperature, waterfrac and enthalpy are three
subarrays within a 'tracers' array that lives in the 'thermal' var struct. We may want to
modify this structure at some point, but for now it seemed better to minimize changes.

I changed some variables names, e.g. from tend_layerThickness to layerThicknessTendency,
to be consistent with MPAS camelCase conventions.

I compared the old and new advection options in a dome problem run for 10 years.
There are small differences associated with temperature advection, but these result
in thickness changes of only a fraction of a meter over 10 years.

I verified with the new scheme that answers on 4 cores (on a Mac) are the same as the
answers on 1 core, within machine roundoff.

It remains to set up an EISMINT-2 test case and do a closer comparison between
MPAS LI and CISM.  If run on the same grid using SIA dynamics and identical input
parameters, the results should be very similar.